Move {circuit,module}_builder.rs
