+++ src/builder/circuit.rs
